refactor to move test to common
